New TSD compiler.
TSD, Template Scanner Compiler, is a compiler that takes a tsd input
file with scanner rules, and compiles it into a efficient template
scanner.

Why?
As it will now be easier to maintain different flavors of the
template scanner.
Also improvements and bugfixes applied to the compiled scanner
will benefit all flavors.

Example:
In zotonic they currently have their own fork of erlydtl, where a bug
fix has been applied to correctly track line numbers within multiline
comments. This bug is still not fixed (maybe not even known about!)
in the main erlydtl project.

TODO:
Document the TSD file format (a little too early still, as it is evolving,
but waiting too long, and the risk of it not being written is obvious).